microservices/user_service/requirements.txt
flask 
mysql-connector-python
